requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
djangorestframework==3.15.1
nose==1.3.7
rounding==0.4
sqlparse==0.5.0
typing-extensions==4.11.0
